FileUtils.forceDelete() throws if the file doesn't exist, contra to expectations of what "force" usually means when deleting.

+++ vassal-app/src/main/java/VASSAL/tools/io/ZipArchive.java
@@ -452,7 +452,9 @@ public class ZipArchive implements FileArchive {
 
     // Replace old archive with temp archive.
     try {
-      FileUtils.forceDelete(archiveFile);
+      if (archiveFile.exists()) {
+        FileUtils.forceDelete(archiveFile);
+      }
       FileUtils.moveFile(tmpFile, archiveFile);
     }
     catch (IOException e) {
